% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_em_functions.R
\name{compensator}
\alias{compensator}
\title{Wrapper for \code{compute_I$Compensator}.}
\usage{
compensator(out, inp, pp_obj, parms, k, t = NULL)
}
\arguments{
\item{out}{index for which component of \code{pp} object is the output process.}

\item{inp}{a vector of indices for which components of \code{pp} object are the input process.}

\item{pp_obj}{a \code{pp} object.}

\item{parms}{parameters of the Hawkes process formatted as described for \code{set_parms}.}

\item{k}{output from \code{set_kernel(kernel_type)}. Currently only implemented for \code{kernel_type = "dgamma"}.}

\item{#}{t: override default time interval \code{[0, attr(events, "end_time"]} with a different upper bound.}
}
\value{
vector of \code{length(inp)}, where the j-th element is the (scalar) value of the compensator for \code{events[[out]]} and \code{events[[inp[j]]]}.
}
\description{
Similar to \code{intensity}. Takes an m-dimensional multivariate process and computes the compensator for indicated output and input processes.
}
