% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_em_functions.R
\name{compute_alpha}
\alias{compute_alpha}
\title{Computes the intensity parameter of the Hawkes process.}
\usage{
compute_alpha(out, inp, pp_obj, parms, k, list_P)
}
\arguments{
\item{out}{index for which component of \code{pp} object is the output process.}

\item{inp}{index for which component of \code{pp} object is the input process.}

\item{pp_obj}{a \code{pp} object.}

\item{k}{output from \code{set_kernel(kernel_type)}. Currently only implemented for \code{kernel_type = "dgamma"}.}

\item{list_P}{output from \code{probs}.}
}
\value{
The (scalar) value of the intensity parameter.
}
\description{
Compute the intensity parameters for single intput/output process combination.
}
