% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_em_functions.R
\name{compute_mu}
\alias{compute_mu}
\title{Computes the baserate parameter of the Hawkes process.}
\usage{
compute_mu(out, inp, pp_obj, list_P)
}
\arguments{
\item{out}{index for which component of \code{pp} object is the output process.}

\item{inp}{a vector of indices for which components of \code{pp} object are the input process.}

\item{pp_obj}{a \code{pp} object.}

\item{list_P}{output from \code{probs}.}
}
\value{
The (scalar) value of the (constant) baserate.
}
\description{
Computes the (constant) baserate for a single output process.
}
