% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_em_functions.R
\name{compute_Q}
\alias{compute_Q}
\title{Computes the Q function the EM algorithm.}
\usage{
compute_Q(out, inp, pp_obj, parms, k, list_P)
}
\arguments{
\item{out}{index for which component of \code{pp} object is the output process.}

\item{inp}{index for which component of \code{pp} object is the input process.}

\item{pp_obj}{a \code{pp} object.}

\item{parms}{parameters of the Hawkes process formatted as described for \code{set_parms}.}

\item{k}{output from \code{set_kernel(kernel_type)}. Currently only implemented for \code{kernel_type = "dgamma"}.}

\item{list_P}{output from \code{probs}.}
}
\value{
The (scalar) value of the Q function.
}
\description{
Computes obejctive function of the M step of the EM algorithm, for a single input / output process combination
}
