% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{computeLoglikelihood}
\alias{computeLoglikelihood}
\title{Serially computed Hawkes process log likelihood and gradient}
\usage{
computeLoglikelihood(
  locations,
  times,
  randomRates,
  parameters,
  locationsGradient = FALSE,
  randomRatesGradient = FALSE,
  randomRatesHessian = FALSE
)
}
\arguments{
\item{locations}{Matrix of spatial locations (nxp).}

\item{times}{Vector of times (length=n).}

\item{randomRates}{Vector of random rates (length=n).}

\item{parameters}{Hawkes process parameters (length=6).}

\item{gradient}{Return gradient (or log likelihood)? Defaults to FALSE.}
}
\value{
Hawkes process log likelihood or its gradient.
}
\description{
Called by \code{hpHawkes::test()} to compare output with parallel implementations.
Slow. Not recommended for use.
}
