% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dcData}
\alias{dcData}
\title{Times and spatial locations for gunfire events in Washington D.C. in 2010}
\format{
A data frame with 5022 rows and 3 variables:
\describe{
  \item{X}{Kilometers from western boundary of Washington D.C. in Eastern direction}
  \item{Y}{Kilometers from southern boundary of Washington D.C. in Northern direction}
  \item{Time}{Minutes from beginning of 2010}
}
}
\source{
\url{http://justicetechlab.org/shotspotter-data/}
}
\usage{
dcData
}
\description{
A dataset containing over 5000 gunfire events in Washington D.C. in 2010
}
\keyword{datasets}
