% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_em_functions.R
\name{EM}
\alias{EM}
\title{Runs EM algorithm for Multivariate Hawkes process.}
\usage{
EM(pp_obj, kernel_type, starts, nstarts, conv, maxit, Trace)
}
\arguments{
\item{pp_obj}{a \code{pp} object.}

\item{kernel_type}{currently only implemented for \code{kernel_type "dgamma"}  (optional).}

\item{starts}{starting values for parameters of the Hawkes process formatted as described for \code{set_parms}. If omitted randomly generated starting values are used.}

\item{nstarts}{How many times to run the EM algorithm (to address non-convex likelihood). Default = 1. Note that if \code{nstarts > 1} then \code{starts} should be omitted, or else each different run of the algorithm will converge to the same value.}

\item{conv}{convergenve criterion of EM algorithm (absolute change in complete data log likelihood on successive steps M steps). Default = 1e-4.}

\item{maxit}{maximum number of iterations of the EM algorithm. Default = 500.}

\item{Trace}{(logical) if \code{True}, some graphical convergence diagnostics are displayed during optimization.}
}
\value{
A named list with components \code{c("Solutions", "History")}. \code{"Solutions"} is a matrix with \code{nstarts} rows, the first columns of which cotains convergence value of the incomplete data loglikelihood for those starting values, with the remaining rows containing the parameters estimates. \code{"History"} is an \code{nstarts}-by-\code{maxit} matrix that contains the convergence history for each set of starting values. Both matrices are ordered from the best to worst solutions.
}
\description{
The only required argument is \code{pp_obj}. Repeated calls are made to \code{optim} with \code{method = "Nelder-Mead"} and its default arguments. Nelder-Mead may attempt non-admissilbe parameter values for the kernel densities of the Hawkes process (e.g., a negative scale/shape parameter for the gamma distribution), in which \code{optim} case will throw a warning message -- but this is still faster than optim's \code{method = "L-BFGS-B"}. To Do: 1) catch \code{optim} warnings and build box constraints into kernels. 2) Implement constraints over response kernels. While the source code has many (too many) lines of code directed at parameter constraints over response functions, this is currently not implemented robustly so the \code{constraints} argument is currently omitted from the function call. Also need to improve \code{Trace = T} output.
}
