% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{engineInitial}
\alias{engineInitial}
\title{Initialize HPH engine object}
\usage{
engineInitial(
  locDists,
  N,
  P,
  times,
  parameters = c(1, 6),
  threads,
  simd,
  gpu,
  single
)
}
\arguments{
\item{locDists}{N by N locations distances matrix.}

\item{N}{Number of locations and size of distance matrix.}

\item{P}{Dimension of locations.}

\item{times}{Vector of times.}

\item{parameters}{Hawkes process parameters (length=6).}

\item{threads}{Number of CPU cores to be used.}

\item{simd}{For CPU implementation: no SIMD (\code{0}), SSE (\code{1}) or AVX (\code{2}).}

\item{gpu}{Which GPU to use? If only 1 available, use \code{gpu=1}. Defaults to \code{0}, no GPU.}

\item{single}{Set \code{single=1} if your GPU does not accommodate doubles.}
}
\value{
MDS engine object.
}
\description{
Takes data, parameters and implementation details and returns HPH engine object. Used within
\code{hpHawkes::hmcsampler()}.
}
