% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_em_functions.R
\name{get_parms}
\alias{get_parms}
\title{Formats output of \code{EM}}
\usage{
get_parms(em_obj, solution = 1)
}
\arguments{
\item{em_obj}{the output of a call to \code{EM}}

\item{solution}{a scalar indicating which row of \code{EM$Solutions} to read.}
}
\value{
A named list of numeric arrays with components \code{c("mu", "alpha", "k_parm1", "k_parm2")} defined as follows. Let n_margins denote the \code{attr(pp_obj, "n_margins")}. Then \code{"mu"} is the n_margins-vector of baseline parameters; \code{"alpha"} is the n_margins-by-n_margins matrix of intensity parameters; \code{"k_parm1"} is the n_margins-by-n_margins matrix containing the first arg to the response kernels; \code{"k_parm2"} is the n_margins-by-n_margins matrix containing the second arg to the response kernels.
}
\description{
Reformatrs \code{EM} output to be compatable with other functions that take the \code{parms} argument. Currently only implemented for \code{"dgamma"}.
}
