% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hph.R
\name{getGradient}
\alias{getGradient}
\title{Helper HPH log likelihood gradient function}
\usage{
getGradient(engine)
}
\arguments{
\item{engine}{An HPH engine object.}
}
\value{
HPH log likelihood gradient.
}
\description{
Takes HPH engine object and returns log likelihood gradient.
}
