% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hph.R
\name{getLogLikelihood}
\alias{getLogLikelihood}
\title{Helper Hawkes process log likelihood function}
\usage{
getLogLikelihood(engine)
}
\arguments{
\item{engine}{An HPH engine object.}
}
\value{
Hawkes process log likelihood
}
\description{
Takes HPH engine object and returns log likelihood.
}
