% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hph.R
\name{getProbsSelfExcite}
\alias{getProbsSelfExcite}
\title{Helper Hawkes process probability self-excitatory function}
\usage{
getProbsSelfExcite(engine)
}
\arguments{
\item{engine}{An HPH engine object.}
}
\value{
vector of probabilities each event is self-excitatory in origin
}
\description{
Takes HPH engine object and returns log likelihood.
}
