% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hph.R
\name{getRandomRatesGradient}
\alias{getRandomRatesGradient}
\title{Helper Hawkes process random rates gradient}
\usage{
getRandomRatesGradient(engine)
}
\arguments{
\item{engine}{An HPH engine object.}
}
\value{
vector of log likelihood gradient wrt random rates.
}
\description{
Takes HPH engine object and returns gradient of log likelihood wrt random rates.
}
