% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hph.R
\name{getRandomRatesHessian}
\alias{getRandomRatesHessian}
\title{Helper Hawkes process random rates Hessian}
\usage{
getRandomRatesHessian(engine)
}
\arguments{
\item{engine}{An HPH engine object.}
}
\value{
vector of log likelihood hessian wrt random rates.
}
\description{
Takes HPH engine object and returns Hessian of log likelihood wrt random rates.
}
