% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_em_functions.R
\name{gof}
\alias{gof}
\title{Goodness of fit plots and KS tests for a Hawkes process.}
\usage{
gof(pp_obj, parms, kernel_type = "dgamma")
}
\arguments{
\item{pp_obj}{a \code{pp} object.}

\item{parms}{parameters of the Hawkes process formatted as described for \code{set_parms}. If missing, output is for homogeneous Poisson process.}

\item{kernel_type}{currently only implemented for \code{"dgamma"}.}
}
\value{
qq plot of residual waiting times with KS test(s) against \code{exp(1)}.
}
\description{
If \code{parms} is missing, fit is reported for homogeneous Poisson process.
}
