% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{hmcsampler}
\alias{hmcsampler}
\title{HMC for Bayesian inference of Hawkes model parameters}
\usage{
hmcsampler(n_iter, burnIn = 0, locDists = NULL, times = NULL,
  params = rep(1, 6), latentDimension = 2, trajectory = 0.2,
  threads = 1, simd = 0, gpu = 0, single = 0)
}
\arguments{
\item{n_iter}{Number of MCMC iterations.}

\item{burnIn}{Number of initial samples to throw away.}

\item{locDists}{Locations distance matrix.}

\item{times}{Observation times.}

\item{params}{Length 6, default 1.}

\item{latentDimension}{Dimension of latent space. Integer ranging from 2 to 8.}

\item{trajectory}{HMC trajectory length.}

\item{threads}{Number of CPU cores to be used.}

\item{simd}{For CPU implementation: no SIMD (\code{0}), SSE (\code{1}) or AVX (\code{2}).}

\item{gpu}{Which GPU to use? If only 1 available, use \code{gpu=1}. Defaults to \code{0}, no GPU.}

\item{single}{Set \code{single=1} if your GPU does not accommodate doubles.}
}
\value{
List containing posterior samples, negative log likelihood values (\code{target}) and time to compute (\code{Time}).
}
\description{
Takes a number of settings and returns posterior samples of parameters
for spatio-temporal Hawkes process model.
}
