% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpHawkes.R
\docType{package}
\name{hpHawkes}
\alias{hpHawkes}
\title{hpHawkes}
\description{
hpHawkes facilitates fast Bayesian inference for Hawkes processes through GPU, multi-core CPU, and SIMD vectorization powered implementations of the Hamiltonian Monte Carlo algorithm.
The package relies on Rcpp.
}
\author{
Marc Suchard and Andrew Holbrook
}
