% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_em_functions.R
\name{intensity}
\alias{intensity}
\title{Wrapper for \code{compute_I$Intensity}.}
\usage{
intensity(out, inp, pp_obj, parms, k, Log = F)
}
\arguments{
\item{out}{index for which component of \code{pp} object is the output process.}

\item{inp}{a vector of indices for which components of \code{pp} object are the input process.}

\item{pp_obj}{a \code{pp} object.}

\item{parms}{parameters of the Hawkes process formatted as described for \code{set_parms}.}

\item{k}{output from \code{set_kernel(kernel_type)}. Currently only implemented for \code{kernel_type = "dgamma"}.}

\item{Log}{logical: return \code{alpha*kernel} or \code{log(alpha) + log(kernel)} ?}
}
\value{
A list of \code{length(inp)} components, where the j-th element is \code{compute_I} for \code{pp_obj[[out]]} and \code{pp_obj[[inp[j]]]}.
}
\description{
Takes an m-dimensional multivariate Hawkes process and computes the intensities for indicated output and input processes. Main difference with \code{compute_I} is that \code{inp} can be a vector.
}
