% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_em_functions.R
\name{logL}
\alias{logL}
\title{Computes the incomplete data log likelihood of the Hawkes process.}
\usage{
logL(out, inp, pp_obj, parms, k)
}
\arguments{
\item{out}{index for which component of \code{pp} object is the output process.}

\item{inp}{a vector of indices for which components of \code{pp} object are the input process.}

\item{pp_obj}{a \code{pp} object.}

\item{parms}{parameters of the Hawkes process formatted as described for \code{set_parms}.}

\item{k}{output from \code{set_kernel(kernel_type)}. Currently only implemented for \code{kernel_type = "dgamma"}.}
}
\value{
The (scalar) value of incomplete data log likelihood for the output margin.
}
\description{
Computes the contribution of a single margin to the incomplete data log likelihood of the Hawkes process.
}
