% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_em_functions.R
\name{plot.pp}
\alias{plot.pp}
\title{Default plot method for \code{pp} objects.}
\usage{
\method{plot}{pp}(pp_obj)
}
\arguments{
\item{pp_obj}{a \code{pp} object.}
}
\value{
A plot.
}
\description{
Cumulative event count plot function. To do: add waiting times plots.
}
