% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_em_functions.R
\name{pp}
\alias{pp}
\title{Creates a point process object (S3).}
\usage{
pp(list_times, n_margins = length(list_times),
  end_time = max(unlist(list_times)))
}
\arguments{
\item{list_times}{a list, each component containing the event times for a margin of the point process.}

\item{n_margins}{if \code{list_times} is missing, used to create an empty \code{pp} object.}

\item{end_time}{if \code{list_times}  is missing, used to create an empty \code{pp} object.}
}
\value{
A \code{pp} object (S3).
}
\description{
A \code{pp} object is just a list whose components are the event times of each margin of a point process, plus some additional attributes. Role is analgous to \code{stats::ts}.
}
