% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_em_functions.R
\name{print.pp}
\alias{print.pp}
\title{Print function for \code{pp} objects.}
\usage{
\method{print}{pp}(pp_obj)
}
\arguments{
\item{pp_obj}{a \code{pp} object.}
}
\description{
Print function for \code{pp} objects.
}
