% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{probability_se}
\alias{probability_se}
\title{Get event specific probabilities self-excitatory}
\usage{
probability_se(
  locations,
  times,
  params,
  backgroundRates,
  threads = 0,
  simd = 0,
  gpu = 0,
  single = 0,
  naive = FALSE
)
}
\arguments{
\item{locations}{Matrix of spatial locations (nxp).}

\item{times}{Vector of times.}

\item{params}{Hawkes process parameters (length=4).}

\item{backgroundRates}{Vector of observation specific background rates.}

\item{threads}{Number of CPU cores to be used.}

\item{simd}{For CPU implementation: no SIMD (\code{0}), SSE (\code{1}) or AVX (\code{2}).}

\item{gpu}{Which GPU to use? If only 1 available, use \code{gpu=1}. Defaults to \code{0}, no GPU.}

\item{single}{Set \code{single=1} if your GPU does not accommodate doubles.}

\item{naive}{Just use naive R implementation (very very slow).}
}
\value{
n vector of probabilities.
}
\description{
Takes locations, times and parameters and returns the probability
that each event is a parent (backround) or a child (self-excitatory).
}
