% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_em_functions.R
\name{probs}
\alias{probs}
\title{Normalizes the output of \code{intensity} to be probabilities.}
\usage{
probs(out, inp, pp_obj, parms, k)
}
\arguments{
\item{out}{index for which component of \code{pp} object is the output process.}

\item{inp}{a vector of indices for which components of \code{pp} object are the input process.}

\item{pp_obj}{a \code{pp} object.}

\item{parms}{parameters of the Hawkes process formatted as described for \code{set_parms}.}

\item{k}{output from \code{set_kernel(kernel_type)}. Currently only implemented for \code{kernel_type = "dgamma"}.}
}
\value{
A list of \code{length(inp)} components, where the j-th element is \code{compute_I} for \code{pp_obj[[out]]} and \code{pp_obj[[inp[j]]]}, normalized over the input processes.
}
\description{
Similar to \code{intensity}, except the intensities are rescaled to probabilities.
}
