% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_em_functions.R
\name{Q_fun}
\alias{Q_fun}
\title{Under construction: Wrapper for \code{compute_Q}.}
\usage{
Q_fun(out_inp, pp_obj, parms, k, list_P)
}
\arguments{
\item{out_inp}{n_margins-by-n_margins binary matrix in which \code{out_inp_ij = 1} denotes the output/input combinations on which to call compute_Q. When only one element of \code{out_inp} is equal is 1, this is equivalent to calling compute_Q directly on the output/input pair denoted by that element.}

\item{pp_obj}{a \code{pp} object.}

\item{parms}{parameters of the Hawkes process formatted as described for \code{set_parms}.}

\item{k}{output from \code{set_kernel(kernel_type)}. Currently only implemented for \code{kernel_type = "dgamma"}.}

\item{list_P}{output from \code{probs}.}
}
\value{
The (scalar) sum of the values of the Q functions.
}
\description{
Returns sum of \code{compute_Q} over different the output/input combinations given by \code{out_inp}. Currently only implemented for constraints over a single output process (i.e., only 1 row of \code{out_inp} can have non-zero entries). Used to apply model constraints on response kernels. Yuck.
}
