% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_em_functions.R
\name{res}
\alias{res}
\title{Computes residual waiting times of a Hawkes process.}
\usage{
res(pp_obj, parms, kernel_type = "dgamma")
}
\arguments{
\item{pp_obj}{a \code{pp} object.}

\item{parms}{parameters of the Hawkes process formatted as described for \code{set_parms}. If missing, output is for homogeneous Poisson process.}

\item{kernel_type}{currently only implemented for \code{"dgamma"}.}
}
\value{
A list of length \code{attr(pp_obj, "n_margins")}, each component a sorted vector of residual interevent times for the corresponding margin.
}
\description{
Residuals are computed using the time change theorem (see e.g., Daley and Verrra-Jones, 2013, sec. 7.4). If \code{parms} is missing, computes residuals for homogeneous Poisson process. To do: Re-write this to work as a default \code{resid} method on \code{EM} objects.
}
