% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_em_functions.R
\name{set_kernel}
\alias{set_kernel}
\title{Under construction: Define response kernel functions}
\usage{
set_kernel(kernel_type = "dgamma")
}
\arguments{
\item{kernel_type}{currently only implemented for \code{"dgamma"}.}
}
\value{
A named list of functions with components \code{c("Density", "Dist", "Quantile")}.
}
\description{
For user-chosen \code{kernel_type} in \code{c("dgamma", "dexp", "dlnorm")}, returns a list of repsponse kernel functions. Currently only implemented for \code{"dgamma"}. The goal is to set up various response kernel choices so that they have unifrom args and returns, and catch inadmissable args passed during M-step optimization.
}
