% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_em_functions.R
\name{set_parms}
\alias{set_parms}
\title{Under construction: Generate random starting values for a Hawkes process.}
\usage{
set_parms(pp_obj, kernel_type = "dgamma", constraints = NULL)
}
\arguments{
\item{pp_obj}{a \code{pp} object.}

\item{kernel_type}{currently only implemented for \code{"dgamma"}.}

\item{constraints}{currently not implemented.}
}
\value{
A named list of numeric arrays with components \code{c("mu", "alpha", "k_parm1", "k_parm2")} defined as follows. Let n_margins denote the \code{attr(pp_obj, "n_margins")}. Then \code{"mu"} is the n_margins-vector of baseline parameters; \code{"alpha"} is the n_margins-by-n_margins matrix of intensity parameters; \code{"k_parm1"} is the n_margins-by-n_margins matrix containing the first arg to the response kernels; \code{"k_parm2"} is the n_margins-by-n_margins matrix containing the second arg to the response kernels.
}
\description{
For a given \code{pp} object with a given \code{kernel_type}, generate random starting values for EM algorithm. Currently only implemented for \code{"dgamma"}. To Do: implement other kernels, generate parameters that follow equality constraints.
}
