% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hph.R
\name{setBackgroundRates}
\alias{setBackgroundRates}
\title{Deliver background rates vector to HPH engine object}
\usage{
setBackgroundRates(engine, data)
}
\arguments{
\item{engine}{HPH engine object.}

\item{data}{Background rates vector.}
}
\value{
HPH engine object.
}
\description{
Helper function delivers background rates vector to HPH engine object.
}
