% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hph.R
\name{setLocDistsData}
\alias{setLocDistsData}
\title{Deliver distance matrix to HPH engine object}
\usage{
setLocDistsData(engine, data)
}
\arguments{
\item{engine}{HPH engine object.}

\item{data}{Distance matrix.}
}
\value{
HPH engine object.
}
\description{
Helper function delivers distance matrix to HPH engine object.
}
