% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hph.R
\name{setParameters}
\alias{setParameters}
\title{Deliver parameters to HPH engine object}
\usage{
setParameters(engine, parameters)
}
\arguments{
\item{engine}{HPH engine object.}

\item{parameters}{of Hawkes process likelihood}
}
\value{
HPH engine object.
}
\description{
Helper function delivers Hawkes process likelihood parameters to HPH engine object.
}
