% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hph.R
\name{setRandomRates}
\alias{setRandomRates}
\title{Deliver random rates vector to HPH engine object}
\usage{
setRandomRates(engine, data)
}
\arguments{
\item{engine}{HPH engine object.}

\item{data}{Random rates vector.}
}
\value{
HPH engine object.
}
\description{
Helper function delivers random rates vector to HPH engine object.
}
