% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hph.R
\name{setTimDiffsData}
\alias{setTimDiffsData}
\title{Deliver times difference matrix to HPH engine object}
\usage{
setTimDiffsData(engine, data)
}
\arguments{
\item{engine}{HPH engine object.}

\item{data}{times difference matrix.}
}
\value{
HPH engine object.
}
\description{
Helper function delivers times difference matrix to HPH engine object.
}
