% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hph.R
\name{setTimesData}
\alias{setTimesData}
\title{Deliver times vector to HPH engine object}
\usage{
setTimesData(engine, data)
}
\arguments{
\item{engine}{HPH engine object.}

\item{data}{Times vector.}
}
\value{
HPH engine object.
}
\description{
Helper function delivers times vector to HPH engine object.
}
