% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{Potential}
\alias{Potential}
\title{Potential function and gradient for HMC}
\usage{
Potential(engine, parameters, lbs)
}
\arguments{
\item{engine}{HPH engine object.}

\item{parameters}{(Exponentiated) spatio-temporal Hawkes process parameters (d=6).}

\item{lbs}{Lower bounds on spatial and temporal lengthscales for background rate.}
}
\value{
Potential or its gradient.
}
\description{
Takes HPH engine object and model parameters. Returns
potential (proportional to log posterior) function or its gradient. Using
truncated normal priors.
}
